license, readme, refact
